+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#347](https://github.com/os2display/display-api-service/pull/347)
+  - Added onFlush listener to handle ManyToMany collection changes for relations checksum propagation.
+  - Added command to refresh relation checksums.
+
 ## [2.6.0] - 2025-12-05
 
 - [#330](https://github.com/os2display/display-api-service/pull/330)